Extract CSipHasher to it's own file in crypto/ directory.
This is a move-only commit with the exception of changes to includes.
